Remove FaciaTestGlobal and replace with scheduling before and after all tests onStart was being called before each test causing long delays if the front didn't load
